+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 7.1
 - Export (Excel) date to YYYY-MM-DD format (ISO) in Date.php & Preferences.php
 - Select Date Format: Add Preferences( 'DATE' ) in User.fnc.php, Preferences.php, Date.php & Side.php
 - Fix SQL error TITLE column limit to 50 characters in GradeLevels.php
+- HTML remove radio buttons (File Attached or Embed Link) in PortalNotes.php & PortalPollsNotes.fnc.php
 
 Changes in 7.0.3
 ----------------